Extract has_element state to your own file
