+++ packages/ember-views/lib/views/states.js
@@ -1,6 +1,7 @@
 require("ember-views/views/states/default");
 require("ember-views/views/states/pre_render");
 require("ember-views/views/states/in_buffer");
+require("ember-views/views/states/has_element");
 require("ember-views/views/states/in_dom");
 require("ember-views/views/states/destroying");
