+++ packages/ember-views/lib/views/states/has_element.js
@@ -1,0 +1,86 @@
+require('ember-views/views/states/default');
+
+/**
+@module ember
+@submodule ember-views
+*/
+
+var get = Ember.get, set = Ember.set;
+
+var hasElement = Ember.View.states.hasElement = Ember.create(Ember.View.states._default);
+
+Ember.merge(hasElement, {
+  $: function(view, sel) {
+    var elem = get(view, 'element');
+    return sel ? Ember.$(sel, elem) : Ember.$(elem);
+  },
+
+  getElement: function(view) {
+    var parent = get(view, 'parentView');
+    if (parent) { parent = get(parent, 'element'); }
+    if (parent) { return view.findElementInParentElement(parent); }
+    return Ember.$("#" + get(view, 'elementId'))[0];
+  },
+
+  setElement: function(view, value) {
+    if (value === null) {
+      view.transitionTo('preRender');
+    } else {
+      throw "You cannot set an element to a non-null value when the element is already in the DOM.";
+    }
+
+    return value;
+  },
+
+  // once the view has been inserted into the DOM, rerendering is
+  // deferred to allow bindings to synchronize.
+  rerender: function(view) {
+    view.triggerRecursively('willClearRender');
+
+    view.clearRenderedChildren();
+
+    view.domManager.replace(view);
+    return view;
+  },
+
+  // once the view is already in the DOM, destroying it removes it
+  // from the DOM, nukes its element, and puts it back into the
+  // preRender state if inDOM.
+
+  destroyElement: function(view) {
+    view._notifyWillDestroyElement();
+    view.domManager.remove(view);
+    set(view, 'element', null);
+    if (view._scheduledInsert) {
+      Ember.run.cancel(view._scheduledInsert);
+      view._scheduledInsert = null;
+    }
+    return view;
+  },
+
+  empty: function(view) {
+    var _childViews = view._childViews, len, idx;
+    if (_childViews) {
+      len = _childViews.length;
+      for (idx = 0; idx < len; idx++) {
+        _childViews[idx]._notifyWillDestroyElement();
+      }
+    }
+    view.domManager.empty(view);
+  },
+
+  // Handle events from `Ember.EventDispatcher`
+  handleEvent: function(view, eventName, evt) {
+    if (view.has(eventName)) {
+      // Handler should be able to re-dispatch events, so we don't
+      // preventDefault or stopPropagation.
+      return view.trigger(eventName, evt);
+    } else {
+      return true; // continue event propagation
+    }
+  },
+
+  invokeObserver: function(target, observer) {
+    observer.call(target);
+  }
+});
